:construction: Merge admin update route into default
